+++ src/dev_tools/makeClass/handle_sdk_ts.cpp
@@ -295,7 +295,6 @@ const char* STR_PATH_FILE =
     "    ether?: boolean,\n"
     "    dollars?: boolean,\n"
     "    raw?: boolean,\n"
-    "    toFile?: boolean,\n"
     "  },\n"
     "  options?: RequestInit,\n"
     ") {\n"